🐞 fix(components/drivers): fix cpu timer in multithreading (#7222)
* 🐞 fix(components/drivers): fix cpu timer in multithreading

* 🎈 perf(components): change double to uint64_t

* 🎈 perf(components): add UL suffix

+++ components/drivers/include/drivers/cputime.h
@@ -16,12 +16,12 @@
 
 struct rt_clock_cputime_ops
 {
-    double (*cputime_getres)(void);
+    uint64_t (*cputime_getres)(void);
     uint64_t (*cputime_gettime)(void);
     int (*cputime_settimeout)(uint64_t tick, void (*timeout)(void *param), void *param);
 };
 
-double clock_cpu_getres(void);
+uint64_t clock_cpu_getres(void);
 uint64_t clock_cpu_gettime(void);
 int clock_cpu_settimeout(uint64_t tick, void (*timeout)(void *param), void *param);
 int clock_cpu_issettimeout(void);
